amb8465: rx timeout to resync telegram (#117)
* rx timeout 2 sec in telegram 

If the reader is out of sync, next meter start bytes will add to the rest of telegram before. Typically are 7 ms between two chunks of received data. If detected a delay of 2 seconds or more between the rx data, then assume it is the first block of new telegram.
This should fix the issue #116.