Sensostar storageNr fix
The sensostar heatmeter has 30 storage positions for storing energy consuption values. Every semi-month a new value is written to the storageNr 2 and the following values are pushed down one position. This applies even when the meter in configured for monthly storage. In this case the telegrams contain values from every second storage position, these are (in the first half of the month) storageNr 2, 4, 6, etc. and in the second half storageNr 3, 4, 5, etc. This fix corrects the storageNr handling in the sensostar driver.