kamheat: support 16bit status fields
We just bought a bunch of Kamstrup Multical 303, the type info as
printed on the device is 303W40276. The supplier mounted one of them in
a reversed direction, which is shown on the display as info code
0004000. It appears that this status field is reported as DIF `0x02`,
VIF `0xFF`, VIFE `0x22`, but the `kamheat` driver currently assumes that
it's sent as a 32bit `difvifkey = 04FF22` (or as `ErrorFlags), and these
are not being sent by our devices.

Even though the fields match (some of) the flags in the 32bit status
word, apparently there's no other way to code this in wmbusmeters than
to introduce another data field (and have it squashed/injected into the
other one, because reporting both `status` and `status16` in case
there's been no actual `status` field would be silly, IMHO).

Bug: https://github.com/wmbusmeters/wmbusmeters/issues/1653
Link: https://www.sge.cz/images/MC303/MULTICAL303-Technical-Description-English.pdf